+++ package.json
@@ -21,7 +21,6 @@
   "homepage": "https://github.com/ByteBuddies/ByteBuddies#readme",
   "dependencies": {
     "dotenv": "^16.0.3",
-    "express": "^4.18.2",
     "pg": "^8.9.0",
     "react": "^18.0.1",
     "react-dom": "^18.0.1",
@@ -33,7 +32,6 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@testing-library/react": "^14.0.0",
-    "@types/bcryptjs": "^2.4.2",
     "@types/cookie-parser": "^1.4.3",
     "@types/cookie-session": "^2.0.44",
     "@types/cors": "^2.8.13",
@@ -43,10 +41,10 @@
     "@types/pg": "^8.6.6",
     "babel-jest": "^29.5.0",
     "babel-loader": "^9.1.2",
-    "bcryptjs": "^2.4.3",
     "concurrently": "7.6.0",
     "css-loader": "^6.7.3",
     "dotenv": "^16.0.3",
+    "express-session": "^1.17.3",
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.5.0",
     "nodemon": "^2.0.20",
